Cast random samples to float64

+++ piglot/parameter.py
@@ -196,7 +196,7 @@ class RealParameter(OptimisableParameter):
         Union[float, np.ndarray]
             Random value for the parameter.
         """
-        samples = self.prior.sample((self.num_components,)).numpy()
+        samples = self.prior.sample((self.num_components,)).to(torch.float64).numpy()
         return np.clip(samples, self.lbound, self.ubound)
 
     def log_prob(self, values: torch.Tensor) -> torch.Tensor:
